requirements.txt
requests
pandas
python-dotenv
xlrd
selenium
tqdm
openpyxl
pymssql
google-cloud-bigquery